+++ packages/dara-core/tests/js/use-action.spec.tsx
@@ -5,10 +5,12 @@ import { useState } from 'react';
 
 import { INPUT, TOGGLE } from '@/actions/update-variable';
 import { clearRegistries_TEST } from '@/shared/interactivity/store';
+import { clearActionHandlerCache_TEST } from '@/shared/utils/use-action';
 
 import { useAction, useVariable } from '../../js/shared';
 import {
     Action,
+    ActionImpl,
     AnnotatedAction,
     DerivedDataVariable,
     DerivedVariable,
@@ -22,10 +24,16 @@ import {
 import { MockWebSocketClient, Wrapper, server, wrappedRender } from './utils';
 
 describe('useAction', () => {
-    beforeEach(() => server.listen({ onUnhandledRequest: 'error' }));
+    beforeEach(() => {
+        server.listen({ onUnhandledRequest: 'error' });
+        window.localStorage.clear();
+        jest.restoreAllMocks();
+
+        clearRegistries_TEST();
+        clearActionHandlerCache_TEST();
+    });
     afterEach(() => {
         server.resetHandlers();
-        clearRegistries_TEST();
     });
     afterAll(() => server.close());
 
@@ -825,5 +833,190 @@ describe('useAction', () => {
         });
 
         await waitFor(() => expect(history.location.search).toBe('?q=default-value'));
+    });
+
+    it('should handle arbitrary ActionImpl', async () => {
+        interface CustomActionImpl extends ActionImpl {
+            extra: number;
+        }
+
+        interface CustomActionImpl2 extends ActionImpl {
+            other: string;
+        }
+
+        const customAction: CustomActionImpl = {
+            __typename: 'ActionImpl',
+            extra: 10,
+            name: 'CustomAction',
+        };
+
+        const customAction2: CustomActionImpl2 = {
+            __typename: 'ActionImpl',
+            name: 'CustomAction2',
+            other: 'other',
+        };
+
+        const onUnhandledAction = jest.fn();
+
+        // Use render hook and pass in a custom action handler
+        const { result } = renderHook(
+            () =>
+                useAction([customAction, customAction2], {
+                    onUnhandledAction,
+                }),
+            { wrapper: ({ children }) => <Wrapper>{children}</Wrapper> }
+        );
+
+        await waitFor(() => {
+            expect(result.current[0]).toBeInstanceOf(Function);
+            expect(result.current[1]).toEqual(false);
+        });
+
+        // Call the action handler
+        await act(async () => {
+            await result.current[0]('input');
+        });
+
+        // There should be two calls to onUnhandledAction
+        expect(onUnhandledAction).toHaveBeenCalledTimes(2);
+
+        // Check that the onUnhandledAction handler was called with the context and the action1
+        expect(onUnhandledAction).toHaveBeenCalledWith(
+            expect.objectContaining({
+                input: 'input',
+            }),
+            customAction
+        );
+
+        // Check it was also called with action2
+        expect(onUnhandledAction).toHaveBeenCalledWith(
+            expect.objectContaining({
+                input: 'input',
+            }),
+            customAction2
+        );
+    });
+
+    it('should handle AnnotatedAction returning multiple custom actions', async () => {
+        interface CustomActionImpl extends ActionImpl {
+            extra: number;
+        }
+
+        interface CustomActionImpl2 extends ActionImpl {
+            other: string;
+        }
+
+        const customAction: CustomActionImpl = {
+            __typename: 'ActionImpl',
+            extra: 10,
+            name: 'CustomAction',
+        };
+
+        const customAction2: CustomActionImpl2 = {
+            __typename: 'ActionImpl',
+            name: 'CustomAction2',
+            other: 'other',
+        };
+
+        const variable: SingleVariable<string> = {
+            __typename: 'Variable',
+            default: 'value',
+            nested: [],
+            uid: 'uid',
+        };
+
+        const annotated: AnnotatedAction = {
+            definition_uid: 'definition',
+            dynamic_kwargs: {
+                foo: variable,
+            },
+            uid: 'uid',
+        };
+
+        const onUnhandledAction = jest.fn();
+
+        const wsClient = new MockWebSocketClient('uid');
+
+        const { result } = renderHook(
+            () =>
+                useAction(annotated, {
+                    onUnhandledAction,
+                }),
+            { wrapper: ({ children }) => <Wrapper client={wsClient}>{children}</Wrapper> }
+        );
+
+        // store received message
+        let serverReceivedMessage: object = null;
+
+        server.use(
+            rest.post('/api/core/action/:uid', async (req, res, ctx) => {
+                serverReceivedMessage = req.body as object;
+                return res(
+                    ctx.json({
+                        execution_id: 'execution_uid',
+                    })
+                );
+            })
+        );
+
+        // Execute action
+        act(() => {
+            result.current[0]('input');
+        });
+
+        expect(result.current[1]).toEqual(true);
+
+        // Wait for server to receive message
+        await waitFor(() => expect(serverReceivedMessage).not.toBeNull());
+
+        // Send custom impls via wsClient
+        act(() => {
+            // get execution id from the received message
+            const executionId = serverReceivedMessage.execution_id;
+            wsClient.receiveMessage({
+                message: {
+                    action: customAction,
+                    uid: executionId,
+                },
+                type: 'message',
+            });
+            wsClient.receiveMessage({
+                message: {
+                    action: customAction2,
+                    uid: executionId,
+                },
+                type: 'message',
+            });
+
+            // send null to indicate end
+            wsClient.receiveMessage({
+                message: {
+                    action: null,
+                    uid: executionId,
+                },
+                type: 'message',
+            });
+        });
+
+        await waitFor(() => expect(result.current[1]).toEqual(false));
+
+        // There should be two calls to onUnhandledAction
+        await waitFor(() => expect(onUnhandledAction).toHaveBeenCalledTimes(2));
+
+        // Check that the onUnhandledAction handler was called with the context and the action1
+        expect(onUnhandledAction).toHaveBeenCalledWith(
+            expect.objectContaining({
+                input: 'input',
+            }),
+            customAction
+        );
+
+        // Check it was also called with action2
+        expect(onUnhandledAction).toHaveBeenCalledWith(
+            expect.objectContaining({
+                input: 'input',
+            }),
+            customAction2
+        );
     });
 });